requirements.txt
fastapi==0.115.7
pydantic==2.10.6
email-validator==2.2.0
pydantic-settings==2.7.1
uvicorn==0.34.0
alembic==1.14.1
asyncpg==0.30.0
SQLAlchemy==2.0.37
aiobotocore==2.19.0
redis==5.2.1
fastapi-cache2==0.2.2
celery==5.4.0
pyjwt==2.8.0
cryptography==42.0.8
bcrypt==4.2.0